+++ pkg/dashboard/server.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 The KubeOne Authors.
+Copyright 2026 The KubeOne Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -21,88 +21,21 @@ package dashboard
 import (
 	"context"
 	"embed"
-	"errors"
 	"fmt"
 	"net/http"
-	"slices"
 	"time"
 
-	"github.com/angelofallars/htmx-go"
-
-	"k8c.io/kubeone/pkg/clusterstatus"
 	"k8c.io/kubeone/pkg/fail"
 	"k8c.io/kubeone/pkg/kubeconfig"
 	"k8c.io/kubeone/pkg/state"
-	clusterv1alpha1 "k8c.io/machine-controller/sdk/apis/cluster/v1alpha1"
 
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/fields"
 	dynclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 //go:embed assets/*
 var assetsFS embed.FS
-
-type dashboardData struct {
-	ControlPlaneNodes  []node
-	WorkerNodes        []node
-	MachineDeployments []machineDeployment
-}
-
-type node struct {
-	Name              string
-	Status            string
-	IsControlPlane    bool
-	Unschedulable     bool
-	LastHeartbeatTime time.Duration
-	Version           string
-	EtcdOK            bool
-	APIServerOK       bool
-}
-
-type machineDeployment struct {
-	Namespace         string
-	Name              string
-	Replicas          int
-	AvailableReplicas int
-	Kubelet           string
-	Age               time.Duration
-	Machines          []machine
-}
-
-type machine struct {
-	Namespace string
-	Name      string
-	Node      string
-	Kubelet   string
-	Address   string
-	Age       time.Duration
-	Deleted   bool
-}
-
-type pod struct {
-	Name      string
-	Namespace string
-	Status    string
-	Ready     string
-	Restarts  int32
-	Age       time.Duration
-}
-
-type nodesResult struct {
-	ControlPlaneNodes []node
-	WorkerNodes       []node
-}
-
-type uiError struct {
-	Message string
-	Code    int
-}
-
-func (e *uiError) Error() string {
-	return e.Message
-}
 
 func Serve(st *state.State, port int) error {
 	if err := kubeconfig.BuildKubernetesClientset(st); err != nil {
@@ -114,7 +47,7 @@ func Serve(st *state.State, port int) error {
 	http.Handle("POST /scale", scaleHandler(st))
 	http.Handle("POST /rollout", rolloutHandler(st))
 	http.Handle("POST /delete-machine", deleteMachineHandler(st))
-	http.Handle("/pods", podsHandler(st))
+	http.Handle("POST /pods", podsHandler(st))
 	http.Handle("POST /delete-pod", deletePodHandler(st))
 
 	st.Logger.Infoln(fmt.Sprintf("Visit http://localhost:%d to access UI", port))
@@ -122,372 +55,8 @@ func Serve(st *state.State, port int) error {
 		Addr:              fmt.Sprintf(":%d", port),
 		ReadHeaderTimeout: 10 * time.Second,
 	}
-	if err := server.ListenAndServe(); err != nil {
-		return err
-	}
 
-	return nil
-}
-
-func httpHandleError(handler func(http.ResponseWriter, *http.Request) error) http.Handler {
-	return http.HandlerFunc(func(wr http.ResponseWriter, req *http.Request) {
-		if err := handler(wr, req); err != nil {
-			if uiErr, ok := errors.AsType[*uiError](err); ok {
-				http.Error(wr, uiErr.Message, uiErr.Code)
-			} else {
-				http.Error(wr, err.Error(), http.StatusInternalServerError)
-			}
-		}
-	})
-}
-
-func dashboardHandler(st *state.State) http.Handler {
-	return httpHandleError(func(wr http.ResponseWriter, req *http.Request) error {
-		dashboardData, err := getDashboardData(st)
-		if err != nil {
-			return err
-		}
-
-		if err = Layout(dashboardData).Render(req.Context(), wr); err != nil {
-			return err
-		}
-
-		return nil
-	})
-}
-
-func getDashboardData(state *state.State) (*dashboardData, error) {
-	nodes, err := getNodes(state)
-	if err != nil {
-		return nil, err
-	}
-
-	machineDeployments, err := getMachineDeployments(state)
-	if err != nil {
-		return nil, err
-	}
-
-	result := dashboardData{
-		ControlPlaneNodes:  nodes.ControlPlaneNodes,
-		WorkerNodes:        nodes.WorkerNodes,
-		MachineDeployments: machineDeployments,
-	}
-
-	return &result, nil
-}
-
-func getNodes(s *state.State) (*nodesResult, error) {
-	nodes := corev1.NodeList{}
-	nodeListOpts := dynclient.ListOptions{}
-	if err := s.DynamicClient.List(s.Context, &nodes, &nodeListOpts); err != nil {
-		return nil, fail.KubeClient(err, "listing nodes")
-	}
-
-	controlPlaneStatus, err := clusterstatus.Fetch(s, false)
-	if err != nil {
-		return nil, err
-	}
-
-	var result nodesResult
-
-	for _, currNode := range nodes.Items {
-		_, isControlPlane := currNode.Labels["node-role.kubernetes.io/control-plane"]
-		lastCondition := currNode.Status.Conditions[len(currNode.Status.Conditions)-1]
-
-		aNode := node{
-			Name:              currNode.Name,
-			IsControlPlane:    isControlPlane,
-			Status:            string(lastCondition.Type),
-			Unschedulable:     currNode.Spec.Unschedulable,
-			LastHeartbeatTime: time.Since(lastCondition.LastHeartbeatTime.Time).Truncate(time.Second),
-			Version:           currNode.Status.NodeInfo.KubeletVersion,
-			EtcdOK:            findNodeEtcd(controlPlaneStatus, currNode.Name),
-			APIServerOK:       findNodeAPIServer(controlPlaneStatus, currNode.Name),
-		}
-
-		if isControlPlane {
-			result.ControlPlaneNodes = append(result.ControlPlaneNodes, aNode)
-		} else {
-			result.WorkerNodes = append(result.WorkerNodes, aNode)
-		}
-	}
-
-	return &result, nil
-}
-
-func getMachineDeployments(state *state.State) ([]machineDeployment, error) {
-	if state.DynamicClient == nil {
-		return nil, fail.NoKubeClient()
-	}
-
-	result := []machineDeployment{}
-	if !state.Cluster.MachineController.Deploy {
-		return result, nil
-	}
-
-	machineDeployments := clusterv1alpha1.MachineDeploymentList{}
-	err := state.DynamicClient.List(
-		state.Context,
-		&machineDeployments,
-		dynclient.InNamespace(metav1.NamespaceSystem),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	for i := range machineDeployments.Items {
-		currMachineDeployment := &machineDeployments.Items[i]
-		machines, err := getMachines(state, currMachineDeployment)
-		if err != nil {
-			return nil, err
-		}
-
-		result = append(
-			result, machineDeployment{
-				Namespace:         currMachineDeployment.Namespace,
-				Name:              currMachineDeployment.Name,
-				Replicas:          int(*currMachineDeployment.Spec.Replicas),
-				AvailableReplicas: int(currMachineDeployment.Status.AvailableReplicas),
-				Kubelet:           currMachineDeployment.Spec.Template.Spec.Versions.Kubelet,
-				Age:               time.Since(currMachineDeployment.CreationTimestamp.Time).Truncate(time.Second),
-				Machines:          machines,
-			},
-		)
-	}
-
-	return result, nil
-}
-
-func getMachines(state *state.State, md *clusterv1alpha1.MachineDeployment) ([]machine, error) {
-	// filter MachineSets owned by the MachineDeployment
-	machineSets := clusterv1alpha1.MachineSetList{}
-	if err := state.DynamicClient.List(state.Context, &machineSets); err != nil {
-		return nil, err
-	}
-
-	filteredMachineSets := []clusterv1alpha1.MachineSet{}
-	for _, currMS := range machineSets.Items {
-		for _, currMSOR := range currMS.OwnerReferences {
-			if md.UID == currMSOR.UID {
-				filteredMachineSets = append(filteredMachineSets, currMS)
-			}
-		}
-	}
-
-	// filter Machines owned by one of the MachineSets owned by the MachineDeployment
-	machines := clusterv1alpha1.MachineList{}
-	if err := state.DynamicClient.List(state.Context, &machines); err != nil {
-		return nil, err
-	}
-
-	filteredMachines := []clusterv1alpha1.Machine{}
-	for _, currMachine := range machines.Items {
-		for _, currMachineOR := range currMachine.OwnerReferences {
-			for _, currMS := range filteredMachineSets {
-				if currMachineOR.UID == currMS.UID {
-					filteredMachines = append(filteredMachines, currMachine)
-				}
-			}
-		}
-	}
-
-	result := []machine{}
-	for i := range filteredMachines {
-		currMachine := filteredMachines[i]
-		address := getExternalIP(&currMachine)
-
-		var nodeName string
-		if noderef := currMachine.Status.NodeRef; noderef != nil {
-			nodeName = noderef.Name
-		}
-
-		result = append(result, machine{
-			Namespace: currMachine.Namespace,
-			Name:      currMachine.Name,
-			Node:      nodeName,
-			Kubelet:   currMachine.Spec.Versions.Kubelet,
-			Address:   address,
-			Age:       time.Since(currMachine.CreationTimestamp.Time).Truncate(time.Second),
-			Deleted:   !currMachine.DeletionTimestamp.IsZero(),
-		})
-	}
-
-	return result, nil
-}
-
-type scaleForm struct {
-	Namespace string
-	Name      string
-	Direction string
-}
-
-func scaleHandler(st *state.State) http.Handler {
-	return httpHandleError(func(wr http.ResponseWriter, req *http.Request) error {
-		if err := req.ParseForm(); err != nil {
-			return err
-		}
-
-		form := scaleForm{
-			Namespace: req.FormValue("namespace"),
-			Name:      req.FormValue("name"),
-			Direction: req.FormValue("direction"),
-		}
-
-		if form.Namespace == "" || form.Name == "" {
-			return &uiError{
-				Message: "namespace and name are required",
-				Code:    http.StatusBadRequest,
-			}
-		}
-
-		md := clusterv1alpha1.MachineDeployment{}
-		key := dynclient.ObjectKey{Namespace: form.Namespace, Name: form.Name}
-		if err := st.DynamicClient.Get(req.Context(), key, &md); err != nil {
-			return fail.KubeClient(err, "getting MachineDeployment")
-		}
-
-		current := int32(0)
-		if md.Spec.Replicas != nil {
-			current = *md.Spec.Replicas
-		}
-
-		patch := dynclient.MergeFrom(md.DeepCopy())
-
-		switch form.Direction {
-		case "up":
-			current++
-		case "down":
-			if current > 0 {
-				current--
-			}
-		default:
-			return &uiError{
-				Message: "direction must be 'up' or 'down'",
-				Code:    http.StatusBadRequest,
-			}
-		}
-
-		md.Spec.Replicas = &current
-		if err := st.DynamicClient.Patch(req.Context(), &md, patch); err != nil {
-			return fail.KubeClient(err, "patching MachineDeployment")
-		}
-
-		if htmx.IsHTMX(req) {
-			data, err := getDashboardData(st)
-			if err != nil {
-				return err
-			}
-
-			return Layout(data).Render(req.Context(), wr)
-		}
-
-		http.Redirect(wr, req, "/", http.StatusSeeOther)
-
-		return nil
-	})
-}
-
-type rolloutForm struct {
-	Namespace string
-	Name      string
-}
-
-func rolloutHandler(st *state.State) http.Handler {
-	return httpHandleError(func(wr http.ResponseWriter, req *http.Request) error {
-		if err := req.ParseForm(); err != nil {
-			return err
-		}
-
-		form := rolloutForm{
-			Namespace: req.FormValue("namespace"),
-			Name:      req.FormValue("name"),
-		}
-
-		if form.Namespace == "" || form.Name == "" {
-			return &uiError{
-				Message: "namespace and name are required",
-				Code:    http.StatusBadRequest,
-			}
-		}
-
-		md := clusterv1alpha1.MachineDeployment{}
-		key := dynclient.ObjectKey{Namespace: form.Namespace, Name: form.Name}
-		if err := st.DynamicClient.Get(req.Context(), key, &md); err != nil {
-			return fail.KubeClient(err, "getting MachineDeployment")
-		}
-
-		patch := dynclient.MergeFrom(md.DeepCopy())
-
-		if md.Spec.Template.Labels == nil {
-			md.Spec.Template.Labels = make(map[string]string)
-		}
-		md.Spec.Template.Labels["forced-restart"] = time.Now().UTC().Format("2006-01-02T15.04.05Z")
-
-		if err := st.DynamicClient.Patch(req.Context(), &md, patch); err != nil {
-			return fail.KubeClient(err, "patching MachineDeployment")
-		}
-
-		if htmx.IsHTMX(req) {
-			data, err := getDashboardData(st)
-			if err != nil {
-				return err
-			}
-
-			return Layout(data).Render(req.Context(), wr)
-		}
-
-		http.Redirect(wr, req, "/", http.StatusSeeOther)
-
-		return nil
-	})
-}
-
-type deleteMachineForm struct {
-	Namespace string
-	Name      string
-}
-
-func deleteMachineHandler(st *state.State) http.Handler {
-	return httpHandleError(func(wr http.ResponseWriter, req *http.Request) error {
-		if err := req.ParseForm(); err != nil {
-			return err
-		}
-
-		form := deleteMachineForm{
-			Namespace: req.FormValue("namespace"),
-			Name:      req.FormValue("name"),
-		}
-
-		if form.Namespace == "" || form.Name == "" {
-			return &uiError{
-				Message: "namespace and name are required",
-				Code:    http.StatusBadRequest,
-			}
-		}
-
-		m := clusterv1alpha1.Machine{}
-		key := dynclient.ObjectKey{Namespace: form.Namespace, Name: form.Name}
-		if err := st.DynamicClient.Get(req.Context(), key, &m); err != nil {
-			return fail.KubeClient(err, "getting Machine")
-		}
-
-		if err := st.DynamicClient.Delete(req.Context(), &m); err != nil {
-			return fail.KubeClient(err, "deleting Machine")
-		}
-
-		if htmx.IsHTMX(req) {
-			data, err := getDashboardData(st)
-			if err != nil {
-				return err
-			}
-
-			return Layout(data).Render(req.Context(), wr)
-		}
-
-		http.Redirect(wr, req, "/", http.StatusSeeOther)
-
-		return nil
-	})
+	return server.ListenAndServe()
 }
 
 func getPodsForNode(ctx context.Context, dynClient dynclient.Client, nodeName string) ([]pod, error) {
@@ -522,95 +91,4 @@ func getPodsForNode(ctx context.Context, dynClient dynclient.Client, nodeName st
 	}
 
 	return pods, nil
-}
-
-func podsHandler(st *state.State) http.Handler {
-	return httpHandleError(func(wr http.ResponseWriter, req *http.Request) error {
-		if req.Method != http.MethodGet {
-			return &uiError{Message: "method not allowed", Code: http.StatusMethodNotAllowed}
-		}
-
-		nodeName := req.URL.Query().Get("node")
-		if nodeName == "" {
-			return &uiError{Message: "node is required", Code: http.StatusBadRequest}
-		}
-
-		pods, err := getPodsForNode(req.Context(), st.DynamicClient, nodeName)
-		if err != nil {
-			return err
-		}
-
-		return NodePodsTable(nodeName, pods).Render(req.Context(), wr)
-	})
-}
-
-type deletePodForm struct {
-	Namespace string
-	Name      string
-	Node      string
-}
-
-func deletePodHandler(st *state.State) http.Handler {
-	return httpHandleError(func(wr http.ResponseWriter, req *http.Request) error {
-		if err := req.ParseForm(); err != nil {
-			return err
-		}
-
-		form := deletePodForm{
-			Namespace: req.FormValue("namespace"),
-			Name:      req.FormValue("name"),
-			Node:      req.FormValue("node"),
-		}
-
-		if form.Namespace == "" || form.Name == "" || form.Node == "" {
-			return &uiError{Message: "namespace, name, and node are required", Code: http.StatusBadRequest}
-		}
-
-		p := corev1.Pod{}
-		key := dynclient.ObjectKey{Namespace: form.Namespace, Name: form.Name}
-		if err := st.DynamicClient.Get(req.Context(), key, &p); err != nil {
-			return fail.KubeClient(err, "getting Pod")
-		}
-
-		if err := st.DynamicClient.Delete(req.Context(), &p); err != nil {
-			return fail.KubeClient(err, "deleting Pod")
-		}
-
-		pods, err := getPodsForNode(req.Context(), st.DynamicClient, form.Node)
-		if err != nil {
-			return err
-		}
-
-		return NodePodsTable(form.Node, pods).Render(req.Context(), wr)
-	})
-}
-
-func findNodeEtcd(nodes []clusterstatus.NodeStatus, search string) bool {
-	for _, cp := range nodes {
-		if cp.NodeName == search {
-			return cp.Etcd
-		}
-	}
-
-	return false
-}
-
-func findNodeAPIServer(nodes []clusterstatus.NodeStatus, search string) bool {
-	for _, cp := range nodes {
-		if cp.NodeName == search {
-			return cp.APIServer
-		}
-	}
-
-	return false
-}
-
-func getExternalIP(machine *clusterv1alpha1.Machine) string {
-	addressIndex := slices.IndexFunc(machine.Status.Addresses, func(a corev1.NodeAddress) bool { return a.Type == "ExternalIP" })
-	if addressIndex >= 0 {
-		return machine.Status.Addresses[addressIndex].Address
-	}
-
-	// TODO what if no external ip address
-	return ""
 }
